+++ hush.js
@@ -31,9 +31,6 @@ var modelOptions = {
 // 3: Define the User model.
 var Hush = db['log'].define('hush', modelDefinition, modelOptions);
 
-console.log(db['log']);
-
-
 // The authentication controller.
 var HushController = {};