+++ server.js
@@ -2,6 +2,9 @@ const app = require("./app");
 const debug = require("debug")("node-angular");
 const http = require("http");
 
+var AuthController = require("./swim-auth-api/controllers/authController");
+var config = require("./config");
+
 const normalizePort = val => {
   var port = parseInt(val, 10);
 
@@ -46,6 +49,14 @@ const onListening = () => {
 
 const port = normalizePort(process.env.PORT || "3000");
 app.set("port", port);
+
+// Check if the user database is empty, if so create the users
+
+console.log('Creating Content Manager...');
+AuthController.createUsers(config.admin.user, config.admin.password, 1);
+
+console.log('Creating Guest User...');
+AuthController.createUsers(config.guest.user, config.guest.password, 0);
 
 const server = http.createServer(app);
 server.on("error", onError);